+++ src/main/java/com/onenth/OneNth/domain/product/repository/itemRepository/ItemImageRepository.java
@@ -1,4 +1,4 @@
-package com.onenth.OneNth.domain.product.repository;
+package com.onenth.OneNth.domain.product.repository.itemRepository;
 
 import com.onenth.OneNth.domain.product.entity.ItemImage;
 import org.springframework.data.jpa.repository.JpaRepository;
