+++ src/main/java/com/onenth/OneNth/domain/product/repository/itemRepository/TagRepository.java
@@ -1,4 +1,4 @@
-package com.onenth.OneNth.domain.product.repository;
+package com.onenth.OneNth.domain.product.repository.itemRepository;
 
 import com.onenth.OneNth.domain.product.entity.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
