+++ src/main/java/com/onenth/OneNth/domain/product/repository/itemRepository/purchase/PurchaseItemRepository.java
@@ -1,4 +1,4 @@
-package com.onenth.OneNth.domain.product.repository;
+package com.onenth.OneNth.domain.product.repository.itemRepository.purchase;
 
 import com.onenth.OneNth.domain.product.entity.PurchaseItem;
 import org.springframework.data.jpa.repository.JpaRepository;
