+++ src/main/java/com/onenth/OneNth/domain/product/repository/itemRepository/purchase/PurchaseItemRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.onenth.OneNth.domain.product.repository;
+package com.onenth.OneNth.domain.product.repository.itemRepository.purchase;
 
 import com.onenth.OneNth.domain.product.entity.PurchaseItem;
 import com.onenth.OneNth.domain.product.entity.QPurchaseItem;
