+++ src/main/java/com/onenth/OneNth/domain/product/repository/itemRepository/sharing/SharingItemRepository.java
@@ -1,4 +1,4 @@
-package com.onenth.OneNth.domain.product.repository;
+package com.onenth.OneNth.domain.product.repository.itemRepository.sharing;
 
 import com.onenth.OneNth.domain.product.entity.SharingItem;
 import org.springframework.data.jpa.repository.JpaRepository;
