+++ src/main/java/com/onenth/OneNth/domain/product/repository/reviewRepository/purchase/PurchaseReviewImageRepository.java
@@ -1,4 +1,4 @@
-package com.onenth.OneNth.domain.product.repository;
+package com.onenth.OneNth.domain.product.repository.reviewRepository.purchase;
 
 import com.onenth.OneNth.domain.product.entity.PurchaseReviewImage;
 import org.springframework.data.jpa.repository.JpaRepository;
