+++ src/main/java/com/onenth/OneNth/domain/product/service/PurchaseItemService.java
@@ -12,9 +12,9 @@ import com.onenth.OneNth.domain.product.entity.enums.ItemCategory;
 import com.onenth.OneNth.domain.product.entity.enums.ItemType;
 import com.onenth.OneNth.domain.product.entity.enums.PurchaseMethod;
 import com.onenth.OneNth.domain.product.entity.enums.Status;
-import com.onenth.OneNth.domain.product.repository.ItemImageRepository;
-import com.onenth.OneNth.domain.product.repository.PurchaseItemRepository;
-import com.onenth.OneNth.domain.product.repository.TagRepository;
+import com.onenth.OneNth.domain.product.repository.itemRepository.ItemImageRepository;
+import com.onenth.OneNth.domain.product.repository.itemRepository.purchase.PurchaseItemRepository;
+import com.onenth.OneNth.domain.product.repository.itemRepository.TagRepository;
 import com.onenth.OneNth.domain.region.entity.Region;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
