+++ src/main/java/com/onenth/OneNth/domain/product/service/ReviewCommandService.java
@@ -6,7 +6,12 @@ import com.onenth.OneNth.domain.product.converter.ReviewConverter;
 import com.onenth.OneNth.domain.product.dto.ReviewRequestDTO;
 import com.onenth.OneNth.domain.product.dto.ReviewResponseDTO;
 import com.onenth.OneNth.domain.product.entity.*;
-import com.onenth.OneNth.domain.product.repository.*;
+import com.onenth.OneNth.domain.product.repository.itemRepository.purchase.PurchaseItemRepository;
+import com.onenth.OneNth.domain.product.repository.itemRepository.sharing.SharingItemRepository;
+import com.onenth.OneNth.domain.product.repository.reviewRepository.purchase.PurchaseReviewImageRepository;
+import com.onenth.OneNth.domain.product.repository.reviewRepository.purchase.PurchaseReviewRepository;
+import com.onenth.OneNth.domain.product.repository.reviewRepository.sharing.SharingReviewImageRepository;
+import com.onenth.OneNth.domain.product.repository.reviewRepository.sharing.SharingReviewRepository;
 import com.onenth.OneNth.global.apiPayload.code.status.ErrorStatus;
 import com.onenth.OneNth.global.apiPayload.exception.handler.MemberHandler;
 import com.onenth.OneNth.global.apiPayload.exception.handler.SharingItemHandler;
